+++ front/__tests__/models/Base.spec.js
@@ -1,6 +1,7 @@
-import Base from '../../client/models/Base'
-import MockDate from 'mockdate'
-import moment from 'moment-timezone'
+import Base        from '../../client/models/Base'
+import ConvertCase from '../../client/utils/ConvertCase'
+import MockDate    from 'mockdate'
+import moment      from 'moment-timezone'
 
 const subjectClass = (params) => {
   const klass = Base(params)
@@ -108,26 +109,6 @@ describe('Base', () => {
     test('presence eventStartsAt in UTC', () => {
       const subject = subjectClass({eventStartsAt: '2017-03-02T09:00:00.000Z'})
       expect(subject.formatDateWith('eventStartsAt').toString()).toBe('2017-03-02T09:00:00Z')
-    })
-  })
-
-  describe('camelKeysOf', () => {
-
-    const camelKeysParams = {
-      name: 'name',
-      eventStartsAt: '2017-03-04 13:00',
-      eventEndsAt: '2017-03-04 17:00'
-    }
-
-    const snakeKeysParams = {
-      name: 'name',
-      event_starts_at: '2017-03-04 13:00',
-      event_ends_at: '2017-03-04 17:00'
-    }
-
-    it('converts snake case to camel case', () => {
-      const subject = subjectClass({})
-      expect(subject.camelKeysOf(snakeKeysParams)).toEqual(camelKeysParams)
     })
   })
 })
