+++ front/client/reducers/Community.js
@@ -1,14 +1,13 @@
 import { handleActions }  from 'redux-actions'
 import Community          from '../models/Community'
+import ConvertCase          from '../utils/ConvertCase'
 
 export const communityInitialState = new Community()
 
 export const communityReducerMap = {
   CREATE_COMMUNITY: {
     next: (state, action) => {
-      return new Community(
-        (new Community).camelKeysOf(action.payload)
-      )
+      return new Community(ConvertCase.camelKeysOf(action.payload))
     },
     throw: (state, action) => {
       return new Community({
